requirements.txt
pyTelegramBotAPI
getuseragent
beautifulsoup4
tendo
instagrapi
selenium
yt-dlp
ffmpeg-python # yt-dlp additionally requires standalone ffmpeg
# for instagrapi
pillow
